requirements.txt
apache-airflow-providers-google==10.19.0
apache-airflow-providers-postgres==5.11.1
dbt==1.0.0.38.4
dbt-bigquery==1.8.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.20.1
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1

